Add Winamp Linux Qt6 Port
- Complete Qt6 demo application with authentic Winamp look and feel
- Main player window (275x116px classic dimensions)
- Playlist window with green-on-black display
- 10-band graphic equalizer with preamp
- Window snapping/docking like original Winamp
- Frameless windows with draggable title bars
- Classic Winamp colors and 3D button styling
- LED-style time display with green glow
- Oscilloscope and spectrum analyzer visualizations
- Full audio playback via Qt Multimedia
- Interactive volume control
- Platform abstraction layer for Linux (types.h, atomics.h)
- Build type configuration for assertions

+++ Src/Wasabi/bfc/platform/linux.h
@@ -140,10 +140,11 @@ typedef Pixmap HICON;
 
 #endif // wnd
 
-typedef int LRESULT;
-typedef int LPARAM;
-typedef int WPARAM;
-typedef int RPARAM;
+// Message types - use 64-bit for pointer compatibility on x86_64
+typedef long LRESULT;
+typedef long LPARAM;
+typedef unsigned long WPARAM;
+typedef unsigned long RPARAM;
 typedef unsigned int TCHAR;
 typedef long long __int64;
 typedef long long LARGE_INTEGER;
